pyproject.toml
[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[tool.poetry]
name = "clip-reid"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
# NOTE: is poetry slow for you?
#  We found the most likely cause is dependencies not publishing metadata to PyPI.
#  To avoid a huge penalty use these minimum versions for packages you need:
#   - boto3>=1.18.58
#   - botocore>=1.21.65
#  -> To confirm your issue: https://sync.hudlnet.com/x/drGBEg
#  -> More details and solutions: https://sync.hudlnet.com/display/RD/qa/questions/2906
python = ">=3.9,<3.11"
pydantic = "^2.5.2"
numpy = "^1.26.2"
hudl-model-runtime = "^1.1.0"
hudl-datasets = "^3.11.0"
hudl-aws = "^1.9.1"
jupyter = "^1.0.0"
exrex = "^0.11.0"
tomli = "^2.0.1"
matplotlib = "^3.8.2"
open-clip-torch = "^2.23.0"
transformers = "^4.35.2"
huggingface-hub = "^0.19.4"
pyarrow = "^14.0.1"
scikit-learn = "^1.3.2"
fiftyone = "^0.23.1"
quilt3 = "^5.4.0"
fire = "^0.5.0"
onnxruntime = "^1.16.3"
onnx = "^1.15.0"
onnxruntime-gpu = "^1.16.3"
voxel51-eta = "^0.12.1"
pandas = "^1.5.3"
albumentations = "^1.3.1"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
isort = "^5.12"
mypy = "^0.910"
py-spy = "^0.3.8"
pylint = "^2.9.6"
pytest = "^6.2.4"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "hudl"
url = "https://hudlaml-690616407375.d.codeartifact.us-east-1.amazonaws.com/pypi/hudlaml/simple/"
priority = "supplemental"

[tool.pylint.config]
disable = "logging-fstring-interpolation, missing-module-docstring"
enable = "unused-wildcard-import, unused-import, unused-variable"
max-line-length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
